backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
frozendict==2.4.6
html5lib==1.1
idna==3.10
lxml==5.3.1
multitasking==0.0.11
mysqlclient==2.2.7
numpy==2.2.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
yfinance==0.2.52
